Fixed messages_service. I think
