+++ db/db.py
@@ -18,9 +18,16 @@ def get_user(user_id):
     u = db.get_user(user_id)
     return u
 
+def get_user_by_doc_id(id):
+    u = db.get_user_by_doc_id(id)
+    return u
+
 def upsert_user(user:User):
     return db.upsert_user(user)
 
 def save_user(user:User):
     return db.save_user(user)
 
+#TODO: Kill with fire. Leaky abstraction
+def get_user_doc_reference(user_id):
+    return db.get_user_doc_reference(user_id)
